Fix decorator for older browsers
If a method is decorated and declares its parameter types, TypeScript
will add reflection metadata (via the polyfill) that depends on the
global availability of the provided type.
When "method(parameter: XYZ)" is used, window.XYZ must exist at runtime.
This lead to problems with bowsers that don't implement DragEvent.